Move utilities directory to support
