+++ opencore_legacy_patcher/efi_builder/build.py
@@ -14,7 +14,7 @@ from datetime import date
 
 from .. import constants
 
-from ..utilities import utilities
+from ..support import utilities
 
 from .networking import (
     wired,
