+++ opencore_legacy_patcher/efi_builder/graphics_audio.py
@@ -12,7 +12,7 @@ from . import support
 
 from .. import constants
 
-from ..utilities import utilities
+from ..support import utilities
 from ..detections import device_probe
 
 from ..datasets import (
