+++ opencore_legacy_patcher/efi_builder/storage.py
@@ -8,7 +8,7 @@ from . import support
 
 from .. import constants
 
-from ..utilities import utilities
+from ..support import utilities
 from ..detections import device_probe
 
 from ..datasets import (
