+++ opencore_legacy_patcher/support/commit_info.py
