+++ opencore_legacy_patcher/support/global_settings.py
