+++ opencore_legacy_patcher/support/integrity_verification.py
