+++ opencore_legacy_patcher/support/utilities.py
