+++ opencore_legacy_patcher/sys_patch/sys_patch_detect.py
@@ -15,7 +15,7 @@ from ..detections import (
     amfi_detect,
     device_probe
 )
-from ..utilities import (
+from ..support import (
     kdk_handler,
     network_handler,
     utilities
