+++ opencore_legacy_patcher/sys_patch/sys_patch_helpers.py
@@ -15,7 +15,7 @@ from .. import constants
 
 from ..datasets import os_data
 
-from ..utilities import (
+from ..support import (
     bplist,
     generate_smbios,
     utilities
