+++ opencore_legacy_patcher/wx_gui/gui_install_oc.py
@@ -10,7 +10,7 @@ import traceback
 from .. import constants
 
 from ..datasets import os_data
-from ..utilities import install
+from ..support import install
 
 from ..wx_gui import (
     gui_main_menu,
