+++ opencore_legacy_patcher/wx_gui/gui_update.py
@@ -18,7 +18,7 @@ from ..wx_gui import (
     gui_download,
     gui_support
 )
-from ..utilities import (
+from ..support import (
     network_handler,
     updates
 )
